@@ -1945,60 +1945,208 @@ return array(
     'Onoi\\MessageReporter\\NullMessageReporter' => $vendorDir . '/onoi/message-reporter/src/NullMessageReporter.php',
     'Onoi\\MessageReporter\\ObservableMessageReporter' => $vendorDir . '/onoi/message-reporter/src/ObservableMessageReporter.php',
     'Onoi\\MessageReporter\\SpyMessageReporter' => $vendorDir . '/onoi/message-reporter/src/SpyMessageReporter.php',
-    'Opis\\JsonSchema\\Exception\\AbstractSchemaException' => $vendorDir . '/opis/json-schema/src/Exception/AbstractSchemaException.php',
-    'Opis\\JsonSchema\\Exception\\DuplicateSchemaException' => $vendorDir . '/opis/json-schema/src/Exception/DuplicateSchemaException.php',
-    'Opis\\JsonSchema\\Exception\\FilterNotFoundException' => $vendorDir . '/opis/json-schema/src/Exception/FilterNotFoundException.php',
-    'Opis\\JsonSchema\\Exception\\InvalidJsonPointerException' => $vendorDir . '/opis/json-schema/src/Exception/InvalidJsonPointerException.php',
-    'Opis\\JsonSchema\\Exception\\InvalidSchemaDraftException' => $vendorDir . '/opis/json-schema/src/Exception/InvalidSchemaDraftException.php',
-    'Opis\\JsonSchema\\Exception\\InvalidSchemaException' => $vendorDir . '/opis/json-schema/src/Exception/InvalidSchemaException.php',
-    'Opis\\JsonSchema\\Exception\\InvalidSchemaIdException' => $vendorDir . '/opis/json-schema/src/Exception/InvalidSchemaIdException.php',
-    'Opis\\JsonSchema\\Exception\\SchemaDraftNotSupportedException' => $vendorDir . '/opis/json-schema/src/Exception/SchemaDraftNotSupportedException.php',
-    'Opis\\JsonSchema\\Exception\\SchemaKeywordException' => $vendorDir . '/opis/json-schema/src/Exception/SchemaKeywordException.php',
-    'Opis\\JsonSchema\\Exception\\SchemaNotFoundException' => $vendorDir . '/opis/json-schema/src/Exception/SchemaNotFoundException.php',
-    'Opis\\JsonSchema\\Exception\\UnknownMediaTypeException' => $vendorDir . '/opis/json-schema/src/Exception/UnknownMediaTypeException.php',
-    'Opis\\JsonSchema\\FilterContainer' => $vendorDir . '/opis/json-schema/src/FilterContainer.php',
-    'Opis\\JsonSchema\\FormatContainer' => $vendorDir . '/opis/json-schema/src/FormatContainer.php',
-    'Opis\\JsonSchema\\Formats\\AbstractFormat' => $vendorDir . '/opis/json-schema/src/Formats/AbstractFormat.php',
-    'Opis\\JsonSchema\\Formats\\Date' => $vendorDir . '/opis/json-schema/src/Formats/Date.php',
-    'Opis\\JsonSchema\\Formats\\DateTime' => $vendorDir . '/opis/json-schema/src/Formats/DateTime.php',
-    'Opis\\JsonSchema\\Formats\\Email' => $vendorDir . '/opis/json-schema/src/Formats/Email.php',
-    'Opis\\JsonSchema\\Formats\\Hostname' => $vendorDir . '/opis/json-schema/src/Formats/Hostname.php',
-    'Opis\\JsonSchema\\Formats\\IPv4' => $vendorDir . '/opis/json-schema/src/Formats/IPv4.php',
-    'Opis\\JsonSchema\\Formats\\IPv6' => $vendorDir . '/opis/json-schema/src/Formats/IPv6.php',
-    'Opis\\JsonSchema\\Formats\\IdnEmail' => $vendorDir . '/opis/json-schema/src/Formats/IdnEmail.php',
-    'Opis\\JsonSchema\\Formats\\IdnHostname' => $vendorDir . '/opis/json-schema/src/Formats/IdnHostname.php',
-    'Opis\\JsonSchema\\Formats\\Iri' => $vendorDir . '/opis/json-schema/src/Formats/Iri.php',
-    'Opis\\JsonSchema\\Formats\\IriReference' => $vendorDir . '/opis/json-schema/src/Formats/IriReference.php',
-    'Opis\\JsonSchema\\Formats\\JsonPointer' => $vendorDir . '/opis/json-schema/src/Formats/JsonPointer.php',
-    'Opis\\JsonSchema\\Formats\\Regex' => $vendorDir . '/opis/json-schema/src/Formats/Regex.php',
-    'Opis\\JsonSchema\\Formats\\RelativeJsonPointer' => $vendorDir . '/opis/json-schema/src/Formats/RelativeJsonPointer.php',
-    'Opis\\JsonSchema\\Formats\\Time' => $vendorDir . '/opis/json-schema/src/Formats/Time.php',
-    'Opis\\JsonSchema\\Formats\\Uri' => $vendorDir . '/opis/json-schema/src/Formats/Uri.php',
-    'Opis\\JsonSchema\\Formats\\UriReference' => $vendorDir . '/opis/json-schema/src/Formats/UriReference.php',
-    'Opis\\JsonSchema\\Formats\\UriTemplate' => $vendorDir . '/opis/json-schema/src/Formats/UriTemplate.php',
-    'Opis\\JsonSchema\\Formats\\Uuid' => $vendorDir . '/opis/json-schema/src/Formats/Uuid.php',
-    'Opis\\JsonSchema\\IFilter' => $vendorDir . '/opis/json-schema/src/IFilter.php',
-    'Opis\\JsonSchema\\IFilterContainer' => $vendorDir . '/opis/json-schema/src/IFilterContainer.php',
-    'Opis\\JsonSchema\\IFormat' => $vendorDir . '/opis/json-schema/src/IFormat.php',
-    'Opis\\JsonSchema\\IFormatContainer' => $vendorDir . '/opis/json-schema/src/IFormatContainer.php',
-    'Opis\\JsonSchema\\IMediaType' => $vendorDir . '/opis/json-schema/src/IMediaType.php',
-    'Opis\\JsonSchema\\IMediaTypeContainer' => $vendorDir . '/opis/json-schema/src/IMediaTypeContainer.php',
-    'Opis\\JsonSchema\\ISchema' => $vendorDir . '/opis/json-schema/src/ISchema.php',
-    'Opis\\JsonSchema\\ISchemaLoader' => $vendorDir . '/opis/json-schema/src/ISchemaLoader.php',
-    'Opis\\JsonSchema\\IValidator' => $vendorDir . '/opis/json-schema/src/IValidator.php',
-    'Opis\\JsonSchema\\IValidatorHelper' => $vendorDir . '/opis/json-schema/src/IValidatorHelper.php',
+    'Opis\\JsonSchema\\CompliantValidator' => $vendorDir . '/opis/json-schema/src/CompliantValidator.php',
+    'Opis\\JsonSchema\\ContentEncoding' => $vendorDir . '/opis/json-schema/src/ContentEncoding.php',
+    'Opis\\JsonSchema\\ContentMediaType' => $vendorDir . '/opis/json-schema/src/ContentMediaType.php',
+    'Opis\\JsonSchema\\Errors\\CustomError' => $vendorDir . '/opis/json-schema/src/Errors/CustomError.php',
+    'Opis\\JsonSchema\\Errors\\ErrorContainer' => $vendorDir . '/opis/json-schema/src/Errors/ErrorContainer.php',
+    'Opis\\JsonSchema\\Errors\\ErrorFormatter' => $vendorDir . '/opis/json-schema/src/Errors/ErrorFormatter.php',
+    'Opis\\JsonSchema\\Errors\\ValidationError' => $vendorDir . '/opis/json-schema/src/Errors/ValidationError.php',
+    'Opis\\JsonSchema\\Exceptions\\DuplicateSchemaIdException' => $vendorDir . '/opis/json-schema/src/Exceptions/DuplicateSchemaIdException.php',
+    'Opis\\JsonSchema\\Exceptions\\InvalidKeywordException' => $vendorDir . '/opis/json-schema/src/Exceptions/InvalidKeywordException.php',
+    'Opis\\JsonSchema\\Exceptions\\InvalidPragmaException' => $vendorDir . '/opis/json-schema/src/Exceptions/InvalidPragmaException.php',
+    'Opis\\JsonSchema\\Exceptions\\ParseException' => $vendorDir . '/opis/json-schema/src/Exceptions/ParseException.php',
+    'Opis\\JsonSchema\\Exceptions\\SchemaException' => $vendorDir . '/opis/json-schema/src/Exceptions/SchemaException.php',
+    'Opis\\JsonSchema\\Exceptions\\UnresolvedContentEncodingException' => $vendorDir . '/opis/json-schema/src/Exceptions/UnresolvedContentEncodingException.php',
+    'Opis\\JsonSchema\\Exceptions\\UnresolvedContentMediaTypeException' => $vendorDir . '/opis/json-schema/src/Exceptions/UnresolvedContentMediaTypeException.php',
+    'Opis\\JsonSchema\\Exceptions\\UnresolvedException' => $vendorDir . '/opis/json-schema/src/Exceptions/UnresolvedException.php',
+    'Opis\\JsonSchema\\Exceptions\\UnresolvedFilterException' => $vendorDir . '/opis/json-schema/src/Exceptions/UnresolvedFilterException.php',
+    'Opis\\JsonSchema\\Exceptions\\UnresolvedReferenceException' => $vendorDir . '/opis/json-schema/src/Exceptions/UnresolvedReferenceException.php',
+    'Opis\\JsonSchema\\Filter' => $vendorDir . '/opis/json-schema/src/Filter.php',
+    'Opis\\JsonSchema\\Filters\\CommonFilters' => $vendorDir . '/opis/json-schema/src/Filters/CommonFilters.php',
+    'Opis\\JsonSchema\\Filters\\DataExistsFilter' => $vendorDir . '/opis/json-schema/src/Filters/DataExistsFilter.php',
+    'Opis\\JsonSchema\\Filters\\DateTimeFilters' => $vendorDir . '/opis/json-schema/src/Filters/DateTimeFilters.php',
+    'Opis\\JsonSchema\\Filters\\FilterExistsFilter' => $vendorDir . '/opis/json-schema/src/Filters/FilterExistsFilter.php',
+    'Opis\\JsonSchema\\Filters\\FormatExistsFilter' => $vendorDir . '/opis/json-schema/src/Filters/FormatExistsFilter.php',
+    'Opis\\JsonSchema\\Filters\\GlobalVarExistsFilter' => $vendorDir . '/opis/json-schema/src/Filters/GlobalVarExistsFilter.php',
+    'Opis\\JsonSchema\\Filters\\SchemaExistsFilter' => $vendorDir . '/opis/json-schema/src/Filters/SchemaExistsFilter.php',
+    'Opis\\JsonSchema\\Filters\\SlotExistsFilter' => $vendorDir . '/opis/json-schema/src/Filters/SlotExistsFilter.php',
+    'Opis\\JsonSchema\\Format' => $vendorDir . '/opis/json-schema/src/Format.php',
+    'Opis\\JsonSchema\\Formats\\DateTimeFormats' => $vendorDir . '/opis/json-schema/src/Formats/DateTimeFormats.php',
+    'Opis\\JsonSchema\\Formats\\IriFormats' => $vendorDir . '/opis/json-schema/src/Formats/IriFormats.php',
+    'Opis\\JsonSchema\\Formats\\MiscFormats' => $vendorDir . '/opis/json-schema/src/Formats/MiscFormats.php',
+    'Opis\\JsonSchema\\Formats\\UriFormats' => $vendorDir . '/opis/json-schema/src/Formats/UriFormats.php',
+    'Opis\\JsonSchema\\Helper' => $vendorDir . '/opis/json-schema/src/Helper.php',
+    'Opis\\JsonSchema\\Info\\DataInfo' => $vendorDir . '/opis/json-schema/src/Info/DataInfo.php',
+    'Opis\\JsonSchema\\Info\\SchemaInfo' => $vendorDir . '/opis/json-schema/src/Info/SchemaInfo.php',
     'Opis\\JsonSchema\\JsonPointer' => $vendorDir . '/opis/json-schema/src/JsonPointer.php',
-    'Opis\\JsonSchema\\Loaders\\File' => $vendorDir . '/opis/json-schema/src/Loaders/File.php',
-    'Opis\\JsonSchema\\Loaders\\Memory' => $vendorDir . '/opis/json-schema/src/Loaders/Memory.php',
-    'Opis\\JsonSchema\\MediaTypeContainer' => $vendorDir . '/opis/json-schema/src/MediaTypeContainer.php',
-    'Opis\\JsonSchema\\MediaTypes\\Json' => $vendorDir . '/opis/json-schema/src/MediaTypes/Json.php',
-    'Opis\\JsonSchema\\MediaTypes\\Text' => $vendorDir . '/opis/json-schema/src/MediaTypes/Text.php',
+    'Opis\\JsonSchema\\Keyword' => $vendorDir . '/opis/json-schema/src/Keyword.php',
+    'Opis\\JsonSchema\\KeywordValidator' => $vendorDir . '/opis/json-schema/src/KeywordValidator.php',
+    'Opis\\JsonSchema\\KeywordValidators\\AbstractKeywordValidator' => $vendorDir . '/opis/json-schema/src/KeywordValidators/AbstractKeywordValidator.php',
+    'Opis\\JsonSchema\\KeywordValidators\\CallbackKeywordValidator' => $vendorDir . '/opis/json-schema/src/KeywordValidators/CallbackKeywordValidator.php',
+    'Opis\\JsonSchema\\KeywordValidators\\PragmaKeywordValidator' => $vendorDir . '/opis/json-schema/src/KeywordValidators/PragmaKeywordValidator.php',
+    'Opis\\JsonSchema\\Keywords\\AbstractRefKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/AbstractRefKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\AdditionalItemsKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/AdditionalItemsKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\AdditionalPropertiesKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/AdditionalPropertiesKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\AllOfKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/AllOfKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\AnyOfKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/AnyOfKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ConstDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ConstDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ConstKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ConstKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ContainsKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ContainsKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ContentEncodingKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ContentEncodingKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ContentMediaTypeKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ContentMediaTypeKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ContentSchemaKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ContentSchemaKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\DefaultKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/DefaultKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\DependenciesKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/DependenciesKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\DependentRequiredKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/DependentRequiredKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\DependentSchemasKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/DependentSchemasKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\EnumDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/EnumDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\EnumKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/EnumKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ErrorTrait' => $vendorDir . '/opis/json-schema/src/Keywords/ErrorTrait.php',
+    'Opis\\JsonSchema\\Keywords\\ExclusiveMaximumDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ExclusiveMaximumDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ExclusiveMaximumKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ExclusiveMaximumKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ExclusiveMinimumDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ExclusiveMinimumDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ExclusiveMinimumKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ExclusiveMinimumKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\FiltersKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/FiltersKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\FormatDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/FormatDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\FormatKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/FormatKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\IfThenElseKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/IfThenElseKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\ItemsKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/ItemsKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\IterableDataValidationTrait' => $vendorDir . '/opis/json-schema/src/Keywords/IterableDataValidationTrait.php',
+    'Opis\\JsonSchema\\Keywords\\MaxItemsDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MaxItemsDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MaxItemsKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MaxItemsKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MaxLengthDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MaxLengthDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MaxLengthKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MaxLengthKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MaxPropertiesDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MaxPropertiesDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MaxPropertiesKeywords' => $vendorDir . '/opis/json-schema/src/Keywords/MaxPropertiesKeywords.php',
+    'Opis\\JsonSchema\\Keywords\\MaximumDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MaximumDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MaximumKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MaximumKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MinItemsDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MinItemsDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MinItemsKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MinItemsKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MinLengthDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MinLengthDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MinLengthKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MinLengthKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MinPropertiesDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MinPropertiesDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MinPropertiesKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MinPropertiesKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MinimumDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MinimumDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MinimumKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MinimumKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MultipleOfDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MultipleOfDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\MultipleOfKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/MultipleOfKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\NotKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/NotKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\OfTrait' => $vendorDir . '/opis/json-schema/src/Keywords/OfTrait.php',
+    'Opis\\JsonSchema\\Keywords\\OneOfKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/OneOfKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\PatternDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/PatternDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\PatternKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/PatternKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\PatternPropertiesKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/PatternPropertiesKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\PointerRefKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/PointerRefKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\PropertiesKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/PropertiesKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\PropertyNamesKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/PropertyNamesKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\RecursiveRefKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/RecursiveRefKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\RequiredDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/RequiredDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\RequiredKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/RequiredKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\SlotsKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/SlotsKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\TemplateRefKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/TemplateRefKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\TypeKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/TypeKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\URIRefKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/URIRefKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\UnevaluatedItemsKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/UnevaluatedItemsKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\UnevaluatedPropertiesKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/UnevaluatedPropertiesKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\UniqueItemsDataKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/UniqueItemsDataKeyword.php',
+    'Opis\\JsonSchema\\Keywords\\UniqueItemsKeyword' => $vendorDir . '/opis/json-schema/src/Keywords/UniqueItemsKeyword.php',
+    'Opis\\JsonSchema\\Parsers\\DataKeywordTrait' => $vendorDir . '/opis/json-schema/src/Parsers/DataKeywordTrait.php',
+    'Opis\\JsonSchema\\Parsers\\DefaultVocabulary' => $vendorDir . '/opis/json-schema/src/Parsers/DefaultVocabulary.php',
+    'Opis\\JsonSchema\\Parsers\\Draft' => $vendorDir . '/opis/json-schema/src/Parsers/Draft.php',
+    'Opis\\JsonSchema\\Parsers\\DraftOptionTrait' => $vendorDir . '/opis/json-schema/src/Parsers/DraftOptionTrait.php',
+    'Opis\\JsonSchema\\Parsers\\Drafts\\Draft06' => $vendorDir . '/opis/json-schema/src/Parsers/Drafts/Draft06.php',
+    'Opis\\JsonSchema\\Parsers\\Drafts\\Draft07' => $vendorDir . '/opis/json-schema/src/Parsers/Drafts/Draft07.php',
+    'Opis\\JsonSchema\\Parsers\\Drafts\\Draft201909' => $vendorDir . '/opis/json-schema/src/Parsers/Drafts/Draft201909.php',
+    'Opis\\JsonSchema\\Parsers\\Drafts\\Draft202012' => $vendorDir . '/opis/json-schema/src/Parsers/Drafts/Draft202012.php',
+    'Opis\\JsonSchema\\Parsers\\KeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/KeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\KeywordParserTrait' => $vendorDir . '/opis/json-schema/src/Parsers/KeywordParserTrait.php',
+    'Opis\\JsonSchema\\Parsers\\KeywordValidatorParser' => $vendorDir . '/opis/json-schema/src/Parsers/KeywordValidatorParser.php',
+    'Opis\\JsonSchema\\Parsers\\KeywordValidators\\PragmaKeywordValidatorParser' => $vendorDir . '/opis/json-schema/src/Parsers/KeywordValidators/PragmaKeywordValidatorParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\AdditionalItemsKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/AdditionalItemsKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\AdditionalPropertiesKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/AdditionalPropertiesKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\AllOfKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/AllOfKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\AnyOfKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/AnyOfKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\ConstKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/ConstKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\ContainsKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/ContainsKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\ContentEncodingKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/ContentEncodingKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\ContentMediaTypeKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/ContentMediaTypeKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\ContentSchemaKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/ContentSchemaKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\DefaultKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/DefaultKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\DependenciesKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/DependenciesKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\DependentRequiredKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/DependentRequiredKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\DependentSchemasKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/DependentSchemasKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\EnumKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/EnumKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\ExclusiveMaximumKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/ExclusiveMaximumKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\ExclusiveMinimumKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/ExclusiveMinimumKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\FiltersKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/FiltersKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\FormatKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/FormatKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\IfThenElseKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/IfThenElseKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\ItemsKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/ItemsKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\MaxItemsKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/MaxItemsKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\MaxLengthKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/MaxLengthKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\MaxPropertiesKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/MaxPropertiesKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\MaximumKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/MaximumKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\MinItemsKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/MinItemsKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\MinLengthKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/MinLengthKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\MinPropertiesKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/MinPropertiesKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\MinimumKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/MinimumKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\MultipleOfKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/MultipleOfKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\NotKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/NotKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\OneOfKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/OneOfKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\PatternKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/PatternKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\PatternPropertiesKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/PatternPropertiesKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\PropertiesKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/PropertiesKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\PropertyNamesKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/PropertyNamesKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\RefKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/RefKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\RequiredKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/RequiredKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\SlotsKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/SlotsKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\TypeKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/TypeKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\UnevaluatedItemsKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/UnevaluatedItemsKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\UnevaluatedPropertiesKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/UnevaluatedPropertiesKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\Keywords\\UniqueItemsKeywordParser' => $vendorDir . '/opis/json-schema/src/Parsers/Keywords/UniqueItemsKeywordParser.php',
+    'Opis\\JsonSchema\\Parsers\\PragmaParser' => $vendorDir . '/opis/json-schema/src/Parsers/PragmaParser.php',
+    'Opis\\JsonSchema\\Parsers\\Pragmas\\CastPragmaParser' => $vendorDir . '/opis/json-schema/src/Parsers/Pragmas/CastPragmaParser.php',
+    'Opis\\JsonSchema\\Parsers\\Pragmas\\GlobalsPragmaParser' => $vendorDir . '/opis/json-schema/src/Parsers/Pragmas/GlobalsPragmaParser.php',
+    'Opis\\JsonSchema\\Parsers\\Pragmas\\MaxErrorsPragmaParser' => $vendorDir . '/opis/json-schema/src/Parsers/Pragmas/MaxErrorsPragmaParser.php',
+    'Opis\\JsonSchema\\Parsers\\Pragmas\\SlotsPragmaParser' => $vendorDir . '/opis/json-schema/src/Parsers/Pragmas/SlotsPragmaParser.php',
+    'Opis\\JsonSchema\\Parsers\\ResolverTrait' => $vendorDir . '/opis/json-schema/src/Parsers/ResolverTrait.php',
+    'Opis\\JsonSchema\\Parsers\\SchemaParser' => $vendorDir . '/opis/json-schema/src/Parsers/SchemaParser.php',
+    'Opis\\JsonSchema\\Parsers\\VariablesTrait' => $vendorDir . '/opis/json-schema/src/Parsers/VariablesTrait.php',
+    'Opis\\JsonSchema\\Parsers\\Vocabulary' => $vendorDir . '/opis/json-schema/src/Parsers/Vocabulary.php',
+    'Opis\\JsonSchema\\Pragma' => $vendorDir . '/opis/json-schema/src/Pragma.php',
+    'Opis\\JsonSchema\\Pragmas\\CastPragma' => $vendorDir . '/opis/json-schema/src/Pragmas/CastPragma.php',
+    'Opis\\JsonSchema\\Pragmas\\GlobalsPragma' => $vendorDir . '/opis/json-schema/src/Pragmas/GlobalsPragma.php',
+    'Opis\\JsonSchema\\Pragmas\\MaxErrorsPragma' => $vendorDir . '/opis/json-schema/src/Pragmas/MaxErrorsPragma.php',
+    'Opis\\JsonSchema\\Pragmas\\SlotsPragma' => $vendorDir . '/opis/json-schema/src/Pragmas/SlotsPragma.php',
+    'Opis\\JsonSchema\\Resolvers\\ContentEncodingResolver' => $vendorDir . '/opis/json-schema/src/Resolvers/ContentEncodingResolver.php',
+    'Opis\\JsonSchema\\Resolvers\\ContentMediaTypeResolver' => $vendorDir . '/opis/json-schema/src/Resolvers/ContentMediaTypeResolver.php',
+    'Opis\\JsonSchema\\Resolvers\\FilterResolver' => $vendorDir . '/opis/json-schema/src/Resolvers/FilterResolver.php',
+    'Opis\\JsonSchema\\Resolvers\\FormatResolver' => $vendorDir . '/opis/json-schema/src/Resolvers/FormatResolver.php',
+    'Opis\\JsonSchema\\Resolvers\\SchemaResolver' => $vendorDir . '/opis/json-schema/src/Resolvers/SchemaResolver.php',
     'Opis\\JsonSchema\\Schema' => $vendorDir . '/opis/json-schema/src/Schema.php',
-    'Opis\\JsonSchema\\URI' => $vendorDir . '/opis/json-schema/src/URI.php',
-    'Opis\\JsonSchema\\ValidationError' => $vendorDir . '/opis/json-schema/src/ValidationError.php',
+    'Opis\\JsonSchema\\SchemaLoader' => $vendorDir . '/opis/json-schema/src/SchemaLoader.php',
+    'Opis\\JsonSchema\\SchemaValidator' => $vendorDir . '/opis/json-schema/src/SchemaValidator.php',
+    'Opis\\JsonSchema\\Schemas\\AbstractSchema' => $vendorDir . '/opis/json-schema/src/Schemas/AbstractSchema.php',
+    'Opis\\JsonSchema\\Schemas\\BooleanSchema' => $vendorDir . '/opis/json-schema/src/Schemas/BooleanSchema.php',
+    'Opis\\JsonSchema\\Schemas\\EmptySchema' => $vendorDir . '/opis/json-schema/src/Schemas/EmptySchema.php',
+    'Opis\\JsonSchema\\Schemas\\ExceptionSchema' => $vendorDir . '/opis/json-schema/src/Schemas/ExceptionSchema.php',
+    'Opis\\JsonSchema\\Schemas\\LazySchema' => $vendorDir . '/opis/json-schema/src/Schemas/LazySchema.php',
+    'Opis\\JsonSchema\\Schemas\\ObjectSchema' => $vendorDir . '/opis/json-schema/src/Schemas/ObjectSchema.php',
+    'Opis\\JsonSchema\\ValidationContext' => $vendorDir . '/opis/json-schema/src/ValidationContext.php',
     'Opis\\JsonSchema\\ValidationResult' => $vendorDir . '/opis/json-schema/src/ValidationResult.php',
     'Opis\\JsonSchema\\Validator' => $vendorDir . '/opis/json-schema/src/Validator.php',
-    'Opis\\JsonSchema\\ValidatorHelper' => $vendorDir . '/opis/json-schema/src/ValidatorHelper.php',
+    'Opis\\JsonSchema\\Variables' => $vendorDir . '/opis/json-schema/src/Variables.php',
+    'Opis\\JsonSchema\\Variables\\RefVariablesContainer' => $vendorDir . '/opis/json-schema/src/Variables/RefVariablesContainer.php',
+    'Opis\\JsonSchema\\Variables\\VariablesContainer' => $vendorDir . '/opis/json-schema/src/Variables/VariablesContainer.php',
+    'Opis\\String\\Exception\\InvalidCodePointException' => $vendorDir . '/opis/string/src/Exception/InvalidCodePointException.php',
+    'Opis\\String\\Exception\\InvalidStringException' => $vendorDir . '/opis/string/src/Exception/InvalidStringException.php',
+    'Opis\\String\\Exception\\UnicodeException' => $vendorDir . '/opis/string/src/Exception/UnicodeException.php',
+    'Opis\\String\\UnicodeString' => $vendorDir . '/opis/string/src/UnicodeString.php',
+    'Opis\\Uri\\Punycode' => $vendorDir . '/opis/uri/src/Punycode.php',
+    'Opis\\Uri\\PunycodeException' => $vendorDir . '/opis/uri/src/PunycodeException.php',
+    'Opis\\Uri\\Uri' => $vendorDir . '/opis/uri/src/Uri.php',
+    'Opis\\Uri\\UriTemplate' => $vendorDir . '/opis/uri/src/UriTemplate.php',
     'Override' => $vendorDir . '/symfony/polyfill-php83/Resources/stubs/Override.php',
     'PEAR' => $vendorDir . '/pear/pear-core-minimal/src/PEAR.php',
     'PEAR_Error' => $vendorDir . '/pear/pear-core-minimal/src/PEAR.php',
