@@ -204,6 +204,8 @@ class ComposerStaticInit_mediawiki_vendor
         ),
         'O' => 
         array (
+            'Opis\\Uri\\' => 9,
+            'Opis\\String\\' => 12,
             'Opis\\JsonSchema\\' => 16,
             'Onoi\\MessageReporter\\' => 21,
         ),
@@ -556,6 +558,14 @@ class ComposerStaticInit_mediawiki_vendor
         'Peast\\' => 
         array (
             0 => __DIR__ . '/..' . '/mck89/peast/lib/Peast',
+        ),
+        'Opis\\Uri\\' => 
+        array (
+            0 => __DIR__ . '/..' . '/opis/uri/src',
+        ),
+        'Opis\\String\\' => 
+        array (
+            0 => __DIR__ . '/..' . '/opis/string/src',
         ),
         'Opis\\JsonSchema\\' => 
         array (
@@ -2762,60 +2772,208 @@ class ComposerStaticInit_mediawiki_vendor
         'Onoi\\MessageReporter\\NullMessageReporter' => __DIR__ . '/..' . '/onoi/message-reporter/src/NullMessageReporter.php',
         'Onoi\\MessageReporter\\ObservableMessageReporter' => __DIR__ . '/..' . '/onoi/message-reporter/src/ObservableMessageReporter.php',
         'Onoi\\MessageReporter\\SpyMessageReporter' => __DIR__ . '/..' . '/onoi/message-reporter/src/SpyMessageReporter.php',
-        'Opis\\JsonSchema\\Exception\\AbstractSchemaException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/AbstractSchemaException.php',
-        'Opis\\JsonSchema\\Exception\\DuplicateSchemaException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/DuplicateSchemaException.php',
-        'Opis\\JsonSchema\\Exception\\FilterNotFoundException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/FilterNotFoundException.php',
-        'Opis\\JsonSchema\\Exception\\InvalidJsonPointerException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/InvalidJsonPointerException.php',
-        'Opis\\JsonSchema\\Exception\\InvalidSchemaDraftException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/InvalidSchemaDraftException.php',
-        'Opis\\JsonSchema\\Exception\\InvalidSchemaException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/InvalidSchemaException.php',
-        'Opis\\JsonSchema\\Exception\\InvalidSchemaIdException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/InvalidSchemaIdException.php',
-        'Opis\\JsonSchema\\Exception\\SchemaDraftNotSupportedException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/SchemaDraftNotSupportedException.php',
-        'Opis\\JsonSchema\\Exception\\SchemaKeywordException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/SchemaKeywordException.php',
-        'Opis\\JsonSchema\\Exception\\SchemaNotFoundException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/SchemaNotFoundException.php',
-        'Opis\\JsonSchema\\Exception\\UnknownMediaTypeException' => __DIR__ . '/..' . '/opis/json-schema/src/Exception/UnknownMediaTypeException.php',
-        'Opis\\JsonSchema\\FilterContainer' => __DIR__ . '/..' . '/opis/json-schema/src/FilterContainer.php',
-        'Opis\\JsonSchema\\FormatContainer' => __DIR__ . '/..' . '/opis/json-schema/src/FormatContainer.php',
-        'Opis\\JsonSchema\\Formats\\AbstractFormat' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/AbstractFormat.php',
-        'Opis\\JsonSchema\\Formats\\Date' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/Date.php',
-        'Opis\\JsonSchema\\Formats\\DateTime' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/DateTime.php',
-        'Opis\\JsonSchema\\Formats\\Email' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/Email.php',
-        'Opis\\JsonSchema\\Formats\\Hostname' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/Hostname.php',
-        'Opis\\JsonSchema\\Formats\\IPv4' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/IPv4.php',
-        'Opis\\JsonSchema\\Formats\\IPv6' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/IPv6.php',
-        'Opis\\JsonSchema\\Formats\\IdnEmail' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/IdnEmail.php',
-        'Opis\\JsonSchema\\Formats\\IdnHostname' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/IdnHostname.php',
-        'Opis\\JsonSchema\\Formats\\Iri' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/Iri.php',
-        'Opis\\JsonSchema\\Formats\\IriReference' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/IriReference.php',
-        'Opis\\JsonSchema\\Formats\\JsonPointer' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/JsonPointer.php',
-        'Opis\\JsonSchema\\Formats\\Regex' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/Regex.php',
-        'Opis\\JsonSchema\\Formats\\RelativeJsonPointer' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/RelativeJsonPointer.php',
-        'Opis\\JsonSchema\\Formats\\Time' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/Time.php',
-        'Opis\\JsonSchema\\Formats\\Uri' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/Uri.php',
-        'Opis\\JsonSchema\\Formats\\UriReference' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/UriReference.php',
-        'Opis\\JsonSchema\\Formats\\UriTemplate' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/UriTemplate.php',
-        'Opis\\JsonSchema\\Formats\\Uuid' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/Uuid.php',
-        'Opis\\JsonSchema\\IFilter' => __DIR__ . '/..' . '/opis/json-schema/src/IFilter.php',
-        'Opis\\JsonSchema\\IFilterContainer' => __DIR__ . '/..' . '/opis/json-schema/src/IFilterContainer.php',
-        'Opis\\JsonSchema\\IFormat' => __DIR__ . '/..' . '/opis/json-schema/src/IFormat.php',
-        'Opis\\JsonSchema\\IFormatContainer' => __DIR__ . '/..' . '/opis/json-schema/src/IFormatContainer.php',
-        'Opis\\JsonSchema\\IMediaType' => __DIR__ . '/..' . '/opis/json-schema/src/IMediaType.php',
-        'Opis\\JsonSchema\\IMediaTypeContainer' => __DIR__ . '/..' . '/opis/json-schema/src/IMediaTypeContainer.php',
-        'Opis\\JsonSchema\\ISchema' => __DIR__ . '/..' . '/opis/json-schema/src/ISchema.php',
-        'Opis\\JsonSchema\\ISchemaLoader' => __DIR__ . '/..' . '/opis/json-schema/src/ISchemaLoader.php',
-        'Opis\\JsonSchema\\IValidator' => __DIR__ . '/..' . '/opis/json-schema/src/IValidator.php',
-        'Opis\\JsonSchema\\IValidatorHelper' => __DIR__ . '/..' . '/opis/json-schema/src/IValidatorHelper.php',
+        'Opis\\JsonSchema\\CompliantValidator' => __DIR__ . '/..' . '/opis/json-schema/src/CompliantValidator.php',
+        'Opis\\JsonSchema\\ContentEncoding' => __DIR__ . '/..' . '/opis/json-schema/src/ContentEncoding.php',
+        'Opis\\JsonSchema\\ContentMediaType' => __DIR__ . '/..' . '/opis/json-schema/src/ContentMediaType.php',
+        'Opis\\JsonSchema\\Errors\\CustomError' => __DIR__ . '/..' . '/opis/json-schema/src/Errors/CustomError.php',
+        'Opis\\JsonSchema\\Errors\\ErrorContainer' => __DIR__ . '/..' . '/opis/json-schema/src/Errors/ErrorContainer.php',
+        'Opis\\JsonSchema\\Errors\\ErrorFormatter' => __DIR__ . '/..' . '/opis/json-schema/src/Errors/ErrorFormatter.php',
+        'Opis\\JsonSchema\\Errors\\ValidationError' => __DIR__ . '/..' . '/opis/json-schema/src/Errors/ValidationError.php',
+        'Opis\\JsonSchema\\Exceptions\\DuplicateSchemaIdException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/DuplicateSchemaIdException.php',
+        'Opis\\JsonSchema\\Exceptions\\InvalidKeywordException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/InvalidKeywordException.php',
+        'Opis\\JsonSchema\\Exceptions\\InvalidPragmaException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/InvalidPragmaException.php',
+        'Opis\\JsonSchema\\Exceptions\\ParseException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/ParseException.php',
+        'Opis\\JsonSchema\\Exceptions\\SchemaException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/SchemaException.php',
+        'Opis\\JsonSchema\\Exceptions\\UnresolvedContentEncodingException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/UnresolvedContentEncodingException.php',
+        'Opis\\JsonSchema\\Exceptions\\UnresolvedContentMediaTypeException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/UnresolvedContentMediaTypeException.php',
+        'Opis\\JsonSchema\\Exceptions\\UnresolvedException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/UnresolvedException.php',
+        'Opis\\JsonSchema\\Exceptions\\UnresolvedFilterException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/UnresolvedFilterException.php',
+        'Opis\\JsonSchema\\Exceptions\\UnresolvedReferenceException' => __DIR__ . '/..' . '/opis/json-schema/src/Exceptions/UnresolvedReferenceException.php',
+        'Opis\\JsonSchema\\Filter' => __DIR__ . '/..' . '/opis/json-schema/src/Filter.php',
+        'Opis\\JsonSchema\\Filters\\CommonFilters' => __DIR__ . '/..' . '/opis/json-schema/src/Filters/CommonFilters.php',
+        'Opis\\JsonSchema\\Filters\\DataExistsFilter' => __DIR__ . '/..' . '/opis/json-schema/src/Filters/DataExistsFilter.php',
+        'Opis\\JsonSchema\\Filters\\DateTimeFilters' => __DIR__ . '/..' . '/opis/json-schema/src/Filters/DateTimeFilters.php',
+        'Opis\\JsonSchema\\Filters\\FilterExistsFilter' => __DIR__ . '/..' . '/opis/json-schema/src/Filters/FilterExistsFilter.php',
+        'Opis\\JsonSchema\\Filters\\FormatExistsFilter' => __DIR__ . '/..' . '/opis/json-schema/src/Filters/FormatExistsFilter.php',
+        'Opis\\JsonSchema\\Filters\\GlobalVarExistsFilter' => __DIR__ . '/..' . '/opis/json-schema/src/Filters/GlobalVarExistsFilter.php',
+        'Opis\\JsonSchema\\Filters\\SchemaExistsFilter' => __DIR__ . '/..' . '/opis/json-schema/src/Filters/SchemaExistsFilter.php',
+        'Opis\\JsonSchema\\Filters\\SlotExistsFilter' => __DIR__ . '/..' . '/opis/json-schema/src/Filters/SlotExistsFilter.php',
+        'Opis\\JsonSchema\\Format' => __DIR__ . '/..' . '/opis/json-schema/src/Format.php',
+        'Opis\\JsonSchema\\Formats\\DateTimeFormats' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/DateTimeFormats.php',
+        'Opis\\JsonSchema\\Formats\\IriFormats' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/IriFormats.php',
+        'Opis\\JsonSchema\\Formats\\MiscFormats' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/MiscFormats.php',
+        'Opis\\JsonSchema\\Formats\\UriFormats' => __DIR__ . '/..' . '/opis/json-schema/src/Formats/UriFormats.php',
+        'Opis\\JsonSchema\\Helper' => __DIR__ . '/..' . '/opis/json-schema/src/Helper.php',
+        'Opis\\JsonSchema\\Info\\DataInfo' => __DIR__ . '/..' . '/opis/json-schema/src/Info/DataInfo.php',
+        'Opis\\JsonSchema\\Info\\SchemaInfo' => __DIR__ . '/..' . '/opis/json-schema/src/Info/SchemaInfo.php',
         'Opis\\JsonSchema\\JsonPointer' => __DIR__ . '/..' . '/opis/json-schema/src/JsonPointer.php',
-        'Opis\\JsonSchema\\Loaders\\File' => __DIR__ . '/..' . '/opis/json-schema/src/Loaders/File.php',
-        'Opis\\JsonSchema\\Loaders\\Memory' => __DIR__ . '/..' . '/opis/json-schema/src/Loaders/Memory.php',
-        'Opis\\JsonSchema\\MediaTypeContainer' => __DIR__ . '/..' . '/opis/json-schema/src/MediaTypeContainer.php',
-        'Opis\\JsonSchema\\MediaTypes\\Json' => __DIR__ . '/..' . '/opis/json-schema/src/MediaTypes/Json.php',
-        'Opis\\JsonSchema\\MediaTypes\\Text' => __DIR__ . '/..' . '/opis/json-schema/src/MediaTypes/Text.php',
+        'Opis\\JsonSchema\\Keyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keyword.php',
+        'Opis\\JsonSchema\\KeywordValidator' => __DIR__ . '/..' . '/opis/json-schema/src/KeywordValidator.php',
+        'Opis\\JsonSchema\\KeywordValidators\\AbstractKeywordValidator' => __DIR__ . '/..' . '/opis/json-schema/src/KeywordValidators/AbstractKeywordValidator.php',
+        'Opis\\JsonSchema\\KeywordValidators\\CallbackKeywordValidator' => __DIR__ . '/..' . '/opis/json-schema/src/KeywordValidators/CallbackKeywordValidator.php',
+        'Opis\\JsonSchema\\KeywordValidators\\PragmaKeywordValidator' => __DIR__ . '/..' . '/opis/json-schema/src/KeywordValidators/PragmaKeywordValidator.php',
+        'Opis\\JsonSchema\\Keywords\\AbstractRefKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/AbstractRefKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\AdditionalItemsKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/AdditionalItemsKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\AdditionalPropertiesKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/AdditionalPropertiesKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\AllOfKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/AllOfKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\AnyOfKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/AnyOfKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ConstDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ConstDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ConstKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ConstKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ContainsKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ContainsKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ContentEncodingKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ContentEncodingKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ContentMediaTypeKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ContentMediaTypeKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ContentSchemaKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ContentSchemaKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\DefaultKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/DefaultKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\DependenciesKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/DependenciesKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\DependentRequiredKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/DependentRequiredKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\DependentSchemasKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/DependentSchemasKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\EnumDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/EnumDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\EnumKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/EnumKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ErrorTrait' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ErrorTrait.php',
+        'Opis\\JsonSchema\\Keywords\\ExclusiveMaximumDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ExclusiveMaximumDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ExclusiveMaximumKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ExclusiveMaximumKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ExclusiveMinimumDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ExclusiveMinimumDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ExclusiveMinimumKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ExclusiveMinimumKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\FiltersKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/FiltersKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\FormatDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/FormatDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\FormatKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/FormatKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\IfThenElseKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/IfThenElseKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\ItemsKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/ItemsKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\IterableDataValidationTrait' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/IterableDataValidationTrait.php',
+        'Opis\\JsonSchema\\Keywords\\MaxItemsDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MaxItemsDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MaxItemsKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MaxItemsKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MaxLengthDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MaxLengthDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MaxLengthKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MaxLengthKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MaxPropertiesDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MaxPropertiesDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MaxPropertiesKeywords' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MaxPropertiesKeywords.php',
+        'Opis\\JsonSchema\\Keywords\\MaximumDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MaximumDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MaximumKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MaximumKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MinItemsDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MinItemsDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MinItemsKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MinItemsKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MinLengthDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MinLengthDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MinLengthKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MinLengthKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MinPropertiesDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MinPropertiesDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MinPropertiesKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MinPropertiesKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MinimumDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MinimumDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MinimumKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MinimumKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MultipleOfDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MultipleOfDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\MultipleOfKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/MultipleOfKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\NotKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/NotKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\OfTrait' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/OfTrait.php',
+        'Opis\\JsonSchema\\Keywords\\OneOfKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/OneOfKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\PatternDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/PatternDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\PatternKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/PatternKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\PatternPropertiesKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/PatternPropertiesKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\PointerRefKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/PointerRefKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\PropertiesKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/PropertiesKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\PropertyNamesKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/PropertyNamesKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\RecursiveRefKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/RecursiveRefKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\RequiredDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/RequiredDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\RequiredKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/RequiredKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\SlotsKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/SlotsKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\TemplateRefKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/TemplateRefKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\TypeKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/TypeKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\URIRefKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/URIRefKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\UnevaluatedItemsKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/UnevaluatedItemsKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\UnevaluatedPropertiesKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/UnevaluatedPropertiesKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\UniqueItemsDataKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/UniqueItemsDataKeyword.php',
+        'Opis\\JsonSchema\\Keywords\\UniqueItemsKeyword' => __DIR__ . '/..' . '/opis/json-schema/src/Keywords/UniqueItemsKeyword.php',
+        'Opis\\JsonSchema\\Parsers\\DataKeywordTrait' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/DataKeywordTrait.php',
+        'Opis\\JsonSchema\\Parsers\\DefaultVocabulary' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/DefaultVocabulary.php',
+        'Opis\\JsonSchema\\Parsers\\Draft' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Draft.php',
+        'Opis\\JsonSchema\\Parsers\\DraftOptionTrait' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/DraftOptionTrait.php',
+        'Opis\\JsonSchema\\Parsers\\Drafts\\Draft06' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Drafts/Draft06.php',
+        'Opis\\JsonSchema\\Parsers\\Drafts\\Draft07' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Drafts/Draft07.php',
+        'Opis\\JsonSchema\\Parsers\\Drafts\\Draft201909' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Drafts/Draft201909.php',
+        'Opis\\JsonSchema\\Parsers\\Drafts\\Draft202012' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Drafts/Draft202012.php',
+        'Opis\\JsonSchema\\Parsers\\KeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/KeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\KeywordParserTrait' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/KeywordParserTrait.php',
+        'Opis\\JsonSchema\\Parsers\\KeywordValidatorParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/KeywordValidatorParser.php',
+        'Opis\\JsonSchema\\Parsers\\KeywordValidators\\PragmaKeywordValidatorParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/KeywordValidators/PragmaKeywordValidatorParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\AdditionalItemsKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/AdditionalItemsKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\AdditionalPropertiesKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/AdditionalPropertiesKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\AllOfKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/AllOfKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\AnyOfKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/AnyOfKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\ConstKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/ConstKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\ContainsKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/ContainsKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\ContentEncodingKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/ContentEncodingKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\ContentMediaTypeKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/ContentMediaTypeKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\ContentSchemaKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/ContentSchemaKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\DefaultKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/DefaultKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\DependenciesKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/DependenciesKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\DependentRequiredKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/DependentRequiredKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\DependentSchemasKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/DependentSchemasKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\EnumKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/EnumKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\ExclusiveMaximumKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/ExclusiveMaximumKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\ExclusiveMinimumKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/ExclusiveMinimumKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\FiltersKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/FiltersKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\FormatKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/FormatKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\IfThenElseKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/IfThenElseKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\ItemsKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/ItemsKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\MaxItemsKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/MaxItemsKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\MaxLengthKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/MaxLengthKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\MaxPropertiesKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/MaxPropertiesKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\MaximumKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/MaximumKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\MinItemsKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/MinItemsKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\MinLengthKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/MinLengthKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\MinPropertiesKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/MinPropertiesKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\MinimumKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/MinimumKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\MultipleOfKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/MultipleOfKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\NotKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/NotKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\OneOfKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/OneOfKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\PatternKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/PatternKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\PatternPropertiesKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/PatternPropertiesKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\PropertiesKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/PropertiesKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\PropertyNamesKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/PropertyNamesKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\RefKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/RefKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\RequiredKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/RequiredKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\SlotsKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/SlotsKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\TypeKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/TypeKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\UnevaluatedItemsKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/UnevaluatedItemsKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\UnevaluatedPropertiesKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/UnevaluatedPropertiesKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\Keywords\\UniqueItemsKeywordParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Keywords/UniqueItemsKeywordParser.php',
+        'Opis\\JsonSchema\\Parsers\\PragmaParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/PragmaParser.php',
+        'Opis\\JsonSchema\\Parsers\\Pragmas\\CastPragmaParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Pragmas/CastPragmaParser.php',
+        'Opis\\JsonSchema\\Parsers\\Pragmas\\GlobalsPragmaParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Pragmas/GlobalsPragmaParser.php',
+        'Opis\\JsonSchema\\Parsers\\Pragmas\\MaxErrorsPragmaParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Pragmas/MaxErrorsPragmaParser.php',
+        'Opis\\JsonSchema\\Parsers\\Pragmas\\SlotsPragmaParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Pragmas/SlotsPragmaParser.php',
+        'Opis\\JsonSchema\\Parsers\\ResolverTrait' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/ResolverTrait.php',
+        'Opis\\JsonSchema\\Parsers\\SchemaParser' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/SchemaParser.php',
+        'Opis\\JsonSchema\\Parsers\\VariablesTrait' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/VariablesTrait.php',
+        'Opis\\JsonSchema\\Parsers\\Vocabulary' => __DIR__ . '/..' . '/opis/json-schema/src/Parsers/Vocabulary.php',
+        'Opis\\JsonSchema\\Pragma' => __DIR__ . '/..' . '/opis/json-schema/src/Pragma.php',
+        'Opis\\JsonSchema\\Pragmas\\CastPragma' => __DIR__ . '/..' . '/opis/json-schema/src/Pragmas/CastPragma.php',
+        'Opis\\JsonSchema\\Pragmas\\GlobalsPragma' => __DIR__ . '/..' . '/opis/json-schema/src/Pragmas/GlobalsPragma.php',
+        'Opis\\JsonSchema\\Pragmas\\MaxErrorsPragma' => __DIR__ . '/..' . '/opis/json-schema/src/Pragmas/MaxErrorsPragma.php',
+        'Opis\\JsonSchema\\Pragmas\\SlotsPragma' => __DIR__ . '/..' . '/opis/json-schema/src/Pragmas/SlotsPragma.php',
+        'Opis\\JsonSchema\\Resolvers\\ContentEncodingResolver' => __DIR__ . '/..' . '/opis/json-schema/src/Resolvers/ContentEncodingResolver.php',
+        'Opis\\JsonSchema\\Resolvers\\ContentMediaTypeResolver' => __DIR__ . '/..' . '/opis/json-schema/src/Resolvers/ContentMediaTypeResolver.php',
+        'Opis\\JsonSchema\\Resolvers\\FilterResolver' => __DIR__ . '/..' . '/opis/json-schema/src/Resolvers/FilterResolver.php',
+        'Opis\\JsonSchema\\Resolvers\\FormatResolver' => __DIR__ . '/..' . '/opis/json-schema/src/Resolvers/FormatResolver.php',
+        'Opis\\JsonSchema\\Resolvers\\SchemaResolver' => __DIR__ . '/..' . '/opis/json-schema/src/Resolvers/SchemaResolver.php',
         'Opis\\JsonSchema\\Schema' => __DIR__ . '/..' . '/opis/json-schema/src/Schema.php',
-        'Opis\\JsonSchema\\URI' => __DIR__ . '/..' . '/opis/json-schema/src/URI.php',
-        'Opis\\JsonSchema\\ValidationError' => __DIR__ . '/..' . '/opis/json-schema/src/ValidationError.php',
+        'Opis\\JsonSchema\\SchemaLoader' => __DIR__ . '/..' . '/opis/json-schema/src/SchemaLoader.php',
+        'Opis\\JsonSchema\\SchemaValidator' => __DIR__ . '/..' . '/opis/json-schema/src/SchemaValidator.php',
+        'Opis\\JsonSchema\\Schemas\\AbstractSchema' => __DIR__ . '/..' . '/opis/json-schema/src/Schemas/AbstractSchema.php',
+        'Opis\\JsonSchema\\Schemas\\BooleanSchema' => __DIR__ . '/..' . '/opis/json-schema/src/Schemas/BooleanSchema.php',
+        'Opis\\JsonSchema\\Schemas\\EmptySchema' => __DIR__ . '/..' . '/opis/json-schema/src/Schemas/EmptySchema.php',
+        'Opis\\JsonSchema\\Schemas\\ExceptionSchema' => __DIR__ . '/..' . '/opis/json-schema/src/Schemas/ExceptionSchema.php',
+        'Opis\\JsonSchema\\Schemas\\LazySchema' => __DIR__ . '/..' . '/opis/json-schema/src/Schemas/LazySchema.php',
+        'Opis\\JsonSchema\\Schemas\\ObjectSchema' => __DIR__ . '/..' . '/opis/json-schema/src/Schemas/ObjectSchema.php',
+        'Opis\\JsonSchema\\ValidationContext' => __DIR__ . '/..' . '/opis/json-schema/src/ValidationContext.php',
         'Opis\\JsonSchema\\ValidationResult' => __DIR__ . '/..' . '/opis/json-schema/src/ValidationResult.php',
         'Opis\\JsonSchema\\Validator' => __DIR__ . '/..' . '/opis/json-schema/src/Validator.php',
-        'Opis\\JsonSchema\\ValidatorHelper' => __DIR__ . '/..' . '/opis/json-schema/src/ValidatorHelper.php',
+        'Opis\\JsonSchema\\Variables' => __DIR__ . '/..' . '/opis/json-schema/src/Variables.php',
+        'Opis\\JsonSchema\\Variables\\RefVariablesContainer' => __DIR__ . '/..' . '/opis/json-schema/src/Variables/RefVariablesContainer.php',
+        'Opis\\JsonSchema\\Variables\\VariablesContainer' => __DIR__ . '/..' . '/opis/json-schema/src/Variables/VariablesContainer.php',
+        'Opis\\String\\Exception\\InvalidCodePointException' => __DIR__ . '/..' . '/opis/string/src/Exception/InvalidCodePointException.php',
+        'Opis\\String\\Exception\\InvalidStringException' => __DIR__ . '/..' . '/opis/string/src/Exception/InvalidStringException.php',
+        'Opis\\String\\Exception\\UnicodeException' => __DIR__ . '/..' . '/opis/string/src/Exception/UnicodeException.php',
+        'Opis\\String\\UnicodeString' => __DIR__ . '/..' . '/opis/string/src/UnicodeString.php',
+        'Opis\\Uri\\Punycode' => __DIR__ . '/..' . '/opis/uri/src/Punycode.php',
+        'Opis\\Uri\\PunycodeException' => __DIR__ . '/..' . '/opis/uri/src/PunycodeException.php',
+        'Opis\\Uri\\Uri' => __DIR__ . '/..' . '/opis/uri/src/Uri.php',
+        'Opis\\Uri\\UriTemplate' => __DIR__ . '/..' . '/opis/uri/src/UriTemplate.php',
         'Override' => __DIR__ . '/..' . '/symfony/polyfill-php83/Resources/stubs/Override.php',
         'PEAR' => __DIR__ . '/..' . '/pear/pear-core-minimal/src/PEAR.php',
         'PEAR_Error' => __DIR__ . '/..' . '/pear/pear-core-minimal/src/PEAR.php',
