@@ -1,9 +1,9 @@
 <?php return array(
     'root' => array(
         'name' => '__root__',
-        'pretty_version' => 'dev-master',
-        'version' => 'dev-master',
-        'reference' => '439278bf79e3e4e76cebf5fe44a216e5caf884ed',
+        'pretty_version' => '1.0.0+no-version-set',
+        'version' => '1.0.0.0',
+        'reference' => null,
         'type' => 'library',
         'install_path' => __DIR__ . '/../',
         'aliases' => array(),
@@ -11,9 +11,9 @@
     ),
     'versions' => array(
         '__root__' => array(
-            'pretty_version' => 'dev-master',
-            'version' => 'dev-master',
-            'reference' => '439278bf79e3e4e76cebf5fe44a216e5caf884ed',
+            'pretty_version' => '1.0.0+no-version-set',
+            'version' => '1.0.0.0',
+            'reference' => null,
             'type' => 'library',
             'install_path' => __DIR__ . '/../',
             'aliases' => array(),
@@ -470,11 +470,29 @@
             'dev_requirement' => false,
         ),
         'opis/json-schema' => array(
-            'pretty_version' => '1.1.0',
-            'version' => '1.1.0.0',
-            'reference' => '39d8072be8657a329a284339994401066406c43b',
+            'pretty_version' => '2.3.0',
+            'version' => '2.3.0.0',
+            'reference' => 'c48df6d7089a45f01e1c82432348f2d5976f9bfb',
             'type' => 'library',
             'install_path' => __DIR__ . '/../opis/json-schema',
+            'aliases' => array(),
+            'dev_requirement' => false,
+        ),
+        'opis/string' => array(
+            'pretty_version' => '2.0.2',
+            'version' => '2.0.2.0',
+            'reference' => 'ba0b9607b9809462b0e28a11e4881a8d77431feb',
+            'type' => 'library',
+            'install_path' => __DIR__ . '/../opis/string',
+            'aliases' => array(),
+            'dev_requirement' => false,
+        ),
+        'opis/uri' => array(
+            'pretty_version' => '1.1.0',
+            'version' => '1.1.0.0',
+            'reference' => '0f3ca49ab1a5e4a6681c286e0b2cc081b93a7d5a',
+            'type' => 'library',
+            'install_path' => __DIR__ . '/../opis/uri',
             'aliases' => array(),
             'dev_requirement' => false,
         ),
